backend/requirements.txt
# 审计作业平台 - 后端依赖
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
email-validator
redis
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart
openpyxl

# 测试依赖
pytest
pytest-asyncio
pytest-cov
hypothesis
fakeredis
aiosqlite
